td-shim: refactor ipl and memory
Refactor ipl and memory by moving memory related code from ipl.rs
into memory.rs.
