+++ rust-tdshim/src/ipl.rs
@@ -2,39 +2,33 @@
 //
 // SPDX-License-Identifier: BSD-2-Clause-Patent
 
-use uefi_pi::hob;
-
 use elf_loader::elf;
 use elf_loader::elf64::ProgramHeader;
 use pe_loader::pe::{self, Section};
+use td_layout::runtime::{TD_PAYLOAD_BASE, TD_PAYLOAD_SIZE};
 
-use crate::memslice;
-use crate::Memory;
+use crate::{memslice, Memory};
 
-const EXTENDED_FUNCTION_INFO: u32 = 0x80000000;
-const VIRT_PHYS_MEM_SIZES: u32 = 0x80000008;
-
-pub const SIZE_4KB: u64 = 0x00001000u64;
-// pub const SIZE_2MB: u64 = 0x00200000u64;
-
-// pub const LOCAL_APIC_MODE_XAPIC: u64 = 0x1;
-// pub const LOCAL_APIC_MODE_X2APIC: u64 = 0x2;
+const SIZE_4KB: u64 = 0x00001000u64;
 
 pub fn efi_size_to_page(size: u64) -> u64 {
-    (size + SIZE_4KB - 1) / SIZE_4KB
+    // It should saturate, but in case...
+    size.saturating_add(SIZE_4KB - 1) / SIZE_4KB
 }
 
 pub fn efi_page_to_size(page: u64) -> u64 {
-    page * SIZE_4KB
+    // It should saturate, but in case...
+    page.saturating_mul(SIZE_4KB) & !(SIZE_4KB - 1)
 }
 
 pub fn find_and_report_entry_point(
     mem: &mut Memory,
     image_buffer: &[u8],
 ) -> Option<(u64, u64, u64)> {
-    // Safe because we are the only user.
+    // Safe because we are the only user in single-thread context.
     let loaded_buffer = unsafe { memslice::get_mem_slice_mut(memslice::SliceType::Payload) };
     let loaded_buffer_slice = loaded_buffer.as_ptr() as u64;
+
     let res = if elf::is_elf(image_buffer) {
         elf::relocate_elf_with_per_program_header(image_buffer, loaded_buffer, |ph| {
             if !ph.is_executable() {
@@ -63,43 +57,25 @@ pub fn find_and_report_entry_point(
     } else {
         return None;
     };
-    log::info!(
-        "image_entry: {:x}, image_base: {:x}, image_size: {:x}\n",
-        res.0,
-        res.1,
-        res.2
-    );
-    Some(res)
-}
 
-/// CpuGetMemorySpaceSize returns the maximum physical memory addressability of the processor.
-pub fn cpu_get_memory_space_size() -> u8 {
-    let cpuid = unsafe { core::arch::x86_64::__cpuid(EXTENDED_FUNCTION_INFO) };
-
-    let size_of_mem_space = if cpuid.eax >= VIRT_PHYS_MEM_SIZES {
-        let cpuid = unsafe { core::arch::x86_64::__cpuid(VIRT_PHYS_MEM_SIZES) };
-        // CPUID.80000008H:EAX[bits 7-0]: the size of the physical address range
-        cpuid.eax as u8
+    let entry = res.0;
+    let base = res.1;
+    let size = res.2;
+    if base < TD_PAYLOAD_BASE as u64
+        || base >= TD_PAYLOAD_BASE + TD_PAYLOAD_SIZE as u64
+        || size > TD_PAYLOAD_SIZE as u64 - (base - TD_PAYLOAD_BASE)
+        || entry < base
+        || entry > base + size
+    {
+        log::error!("invalid payload binary");
+        None
     } else {
-        // fallback value according to edk2 core
-        36
-    };
-
-    log::info!(
-        "Maximum physical memory addressability of the processor - {}\n",
-        size_of_mem_space
-    );
-
-    // TBD: Currently we only map the 64GB memory, change back to size_of_mem_space once page table
-    // allocator can be ready.
-    core::cmp::min(36, size_of_mem_space)
-}
-
-pub fn get_memory_size(hob: &[u8]) -> u64 {
-    let cpu_men_space_size = cpu_get_memory_space_size() as u32;
-    let cpu_memory_size = 2u64.pow(cpu_men_space_size);
-    let hob_memory_size = hob::get_total_memory_top(hob).unwrap();
-    let mem_size = core::cmp::min(cpu_memory_size, hob_memory_size);
-    log::info!("memory_size: 0x{:x}\n", mem_size);
-    mem_size
+        log::info!(
+            "image_entry: {:x}, image_base: {:x}, image_size: {:x}\n",
+            entry,
+            base,
+            size
+        );
+        Some(res)
+    }
 }
